User groups improvement suggestions from @arj03
