+++ browser.html
@@ -13,6 +13,7 @@
     <link rel="stylesheet" type="text/css" href="css/public.css">
     <link rel="stylesheet" type="text/css" href="css/private.css">
     <link rel="stylesheet" type="text/css" href="css/channels.css">
+    <link rel="stylesheet" type="text/css" href="css/groups.css">
     <link rel="stylesheet" type="text/css" href="css/profile.css">
     <link rel="stylesheet" type="text/css" href="css/connections.css">
     <link rel="stylesheet" type="text/css" href="css/modal.css">
@@ -32,9 +33,10 @@
         <router-link to="/public">{{ $t('pages.public') }}</router-link><new-public-messages></new-public-messages> |
         <router-link to="/private">{{ $t('pages.private') }}</router-link><new-private-messages></new-private-messages> |
         <router-link to="/channels">{{ $t('pages.channels') }}</router-link> |
-        <router-link to="/notifications">{{ $t('pages.notifications') }}</router-link> |
+        <router-link to="/groups">{{ $t('pages.groups') }}</router-link> |
         <router-link :to="{ name: 'profile', params: { feedId: SSB.net.id }}">{{ $t('pages.profile') }}</router-link> |
         <router-link to="/connections">{{ $t('pages.connections') }}</router-link><connected></connected> |
+        <router-link to="/notifications" :title="$t('pages.notifications')">&#128276;</router-link> |
         <router-link to="/settings" :title="$t('pages.settings')">&#9881;</router-link>
         <input type="text" placeholder="thread msg id / feedId" id="goTo" v-model="goToTargetText" v-on:keyup.enter="goToTarget">
       </p>
