--- css/channels.css
@@ -1,5 +1,0 @@
-#channels > div > input {
-  padding: 5px;
-  margin: 0 0px 10px 0px;
-  width: 13rem;
-}
