+++ css/groups.css
@@ -1,0 +1,30 @@
+#groups > div > input {
+  padding: 5px;
+  margin: 0 0px 10px 0px;
+  width: 13rem;
+}
+
+#groups ul.groupMembers
+{
+  list-style: none;
+  margin: 0;
+  padding: 0;
+}
+
+#groups ul.groupMembers li > a > img
+{
+  width: 50px;
+  height: 50px;
+}
+
+#groups ul.groupMembers li
+{
+  float: left;
+  padding-right: 0.5rem;
+}
+
+#groups .clearingDiv
+{
+  clear: both;
+}
+
