+++ ui/browser.js
@@ -42,6 +42,7 @@ require('ssb-browser-core/core').init("/.ssb-lite", optionsForCore);
     const Channel = require('./channel')()
     const Group = require('./group')()
     const Channels = require('./channels')()
+    const Groups = require('./groups')()
     const Thread = require('./thread')()
     const Private = require('./private')(componentsState)
     const Connections = require('./connections')()
@@ -54,6 +55,7 @@ require('ssb-browser-core/core').init("/.ssb-lite", optionsForCore);
     const routes = [
       { name: 'public', path: '/public', component: Public },
       { name: 'channels', path: '/channels', component: Channels },
+      { name: 'groups', path: '/groups', component: Groups },
       { name: 'channel', path: '/channel/:channel', component: Channel, props: true },
       { name: 'group', path: '/group/:group', component: Group, props: true },
       { name: 'thread', path: '/thread/:rootId', component: Thread, props: true },
